Project/MOOD CLASSIFICATION/requirements.txt
streamlit
tensorflow
pillow
opencv-python
numpy
matplotlib
